requirements.txt
Django==3.0
django-allauth==0.42.0
django-crispy-forms==1.8.1
django-tinymce4-lite==1.8.0
django-session-timeout==0.1.0
django-extensions==3.0.5
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
Pillow==7.0.0
django-filter==2.2.0
Unidecode==1.1.1
django-axes==5.6.2
xhtml2pdf==0.2.4
django-simple-history==2.11.0 # https://django-simple-history.readthedocs.io/en/latest/index.html
requests-toolbelt==0.9.1
#djangorestframework==3.11.0
